requirements.txt
aiogram==3.10.0
httpx==0.27.0
httpx-socks==0.9.1
python-dotenv==1.0.1
fake-useragent==1.5.1
socksio==1.0.0
